Make `ResultSearchTextField` private
